requirements.txt
aerich==0.6.3
aiofiles==0.8.0
aiogram==3.0.0b3
aiohttp==3.8.1
aioredis==2.0.1
aiosignal==1.2.0
aiosqlite==0.17.0
anyio==3.6.1
asgiref==3.5.2
asttokens==2.0.5
async-timeout==4.0.2
asyncpg==0.26.0
attrs==21.4.0
Babel==2.9.1
backcall==0.2.0
backports.cached-property==1.0.2
bcrypt==3.2.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
git+https://github.com/Worrik/webapp-shop-models
decorator==5.1.1
dictdiffer==0.9.0
environs==9.5.0
executing==0.8.3
fastapi==0.78.0
fastapi-admin==1.0.4
frozenlist==1.3.0
graphql-core==3.2.1
h11==0.13.0
httptools==0.4.0
idna==3.3
ipython==8.4.0
iso8601==1.0.2
jedi==0.18.1
Jinja2==3.1.2
magic-filter==1.0.7
MarkupSafe==2.1.1
marshmallow==3.16.0
matplotlib-inline==0.1.3
multidict==6.0.2
packaging==21.3
parso==0.8.3
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.30
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==1.9.1
Pygments==2.12.0
pyparsing==3.0.9
pypika-tortoise==0.1.5
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
pytz==2022.1
pytzdata==2020.1
PyYAML==6.0
six==1.16.0
sniffio==1.2.0
stack-data==0.3.0
starlette==0.19.1
strawberry-graphql==0.114.5
toml==0.10.2
tomlkit==0.11.0
tortoise-orm==0.19.1
traitlets==5.3.0
typing-extensions==4.2.0
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
wcwidth==0.2.5
websockets==10.3
yarl==1.7.2
